requirements.txt
Django==3.2.12
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==4.1.1
